+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 71.1.0
 
 * Removed `LinkCheckerApi#upsert_resource_monitor` method as it’s not supported by Link Checker API itself.
 * Add optional `level_of_authentication` parameter to Account API `get_sign_in_url`.